requirements.txt
appdirs==1.4.3
astroid==2.4.1
attrs==19.3.0
black==19.10b0
click==7.1.2
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.4
pathspec==0.8.0
Pillow==7.1.2
pylint==2.5.2
regex==2020.5.7
six==1.14.0
toml==0.10.0
torch==1.5.0
torchvision==0.6.0
typed-ast==1.4.1
wrapt==1.12.1
